refactor(preview-renderer): unify iframe scroll-sync into one code path
Lift the scrollToLine / getScrollRatio handle bodies into scrollSyncDom as
scrollIframeToLine / getIframeScrollRatio; MorphIframe and Q2PreviewIframe now
delegate to the same helpers instead of duplicating them.

Also gate MorphIframe's scroll/click/selectionchange listeners on the srcdoc
load (documentReady), so they bind to the settled document rather than the
pre-load one — restoring HTML preview→editor sync now that the handlers are
stable (and no longer re-attach on every render).
